server/entity: Group the shared combat calculations
